pyproject.toml
[project]
name = "crongle"
version = "0.1.0"
description = ""
authors = [
    {name = "kolikaran1992",email = "kolikaran1992@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "dynalog @ git+ssh://git@github.com/kolikaran1992/dynalog.git",
    "kaggle (>=1.7.4.5,<2.0.0.0)",
    "python-crontab (>=3.3.0,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipykernel = "^7.0.1"

